ColumnProperty extension tests
