+++ src/Migrator/Framework/ColumnPropertyExtensions.cs
@@ -6,12 +6,12 @@ public static class ColumnPropertyExtensions
 {
     public static bool IsSet(this ColumnProperty columnProperty, ColumnProperty flags)
     {
-        return (columnProperty & flags) == flags;
+        return flags != ColumnProperty.None && (columnProperty & flags) == flags;
     }
 
     public static bool IsNotSet(this ColumnProperty columnProperty, ColumnProperty flags)
     {
-        return (columnProperty & (~flags)) == 0;
+        return (columnProperty & flags) == 0;
     }
 
     public static ColumnProperty Set(this ColumnProperty columnProperty, ColumnProperty flags)
